% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajusta_categoricas.R
\name{ajusta_categoricas}
\alias{ajusta_categoricas}
\title{ajusta_categoria}
\usage{
ajusta_categoricas(base)
}
\arguments{
\item{base}{dataset with the variables to adjust}
}
\description{
Função que corrige categoricas entre colchetes
}
\examples{
ajusta_categorica()
}
