% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajusta_extras.R
\name{ajusta_extras}
\alias{ajusta_extras}
\title{Ajusta Extras}
\usage{
ajusta_extras(base, vetor = NULL)
}
\arguments{
\item{base}{Base de dados a ser manipulada}

\item{vetor}{Vetor adicional}
}
\description{
Função para corrigir as questões extras com o padrão antigo
}
\examples{
ajusta_extras()
}
