% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrige_json.R
\name{corrige_json_df}
\alias{corrige_json_df}
\title{Corrige o JSON transformado em pela função transforma_json}
\usage{
corrige_json_df(
  base,
  questionItem = "questionItem",
  answer = "answer",
  extraQuestion = ".extraQuestion",
  varName = "ed_q",
  varNameExtra = "_extra",
  na = "NA",
  keepExtra = NULL
)
}
\arguments{
\item{base}{Base convertida pela função transforma_json}

\item{questionItem}{parte do nome da coluna que indica a questão do questionário. Padrão é questionItem}

\item{answer}{parte do nome da coluna que indica a resposta da questão. Padrão é extra.}

\item{extraQuestion}{parte do nome da coluna que indica resposta extra da questão. Padrão é .extraQuestion}

\item{varName}{nome que será dado para as colunas novas. Padrão é ed_q}

\item{varNameExtra}{nome que irá indicar que aquela coluna representa a resposta da questão extra de determinado item do questionário. Padrão é _extra}

\item{na}{indica qual a marcação de NA Padrão é 'NA'}

\item{keepExtra}{um vetor com os números das questões que tem pergunta extra. Padrão é NULL}
}
\value{
Dataframe corrigido
}
\description{
Essa função corrige algumas marcações originais do JSON transformados,
mantendo somente as informações que indicam qual a pergunta do questionário
original e qual suas informações.
}
