% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cookbook.R
\name{create_cookbook}
\alias{create_cookbook}
\title{Create Cookbook}
\usage{
create_cookbook(dados)
}
\arguments{
\item{dados}{Base de dados a ser utilizada}
}
\description{
Função para criar uma estrutura de dicionário para a base de dados oferecida.
}
\examples{
create_cookbook()
}
