% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummies.R
\name{create_dummies}
\alias{create_dummies}
\title{Create Dummies}
\usage{
create_dummies(dados, variavel, n_opcoes)
}
\arguments{
\item{dados}{Base de dados a ser utilizada}

\item{variavel}{Nome da variável como characther}

\item{n_opcoes}{Número de opções da questão que originou a variável}
}
\description{
Função para criar as variáveis dummies a partir da variável original.
}
\examples{
create_dummies()
}
