% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclui_variaveis.R
\name{exclui_variaveis}
\alias{exclui_variaveis}
\title{Exclui Variáveis}
\usage{
exclui_variaveis(base, vetor_colunas)
}
\arguments{
\item{base}{Base de dados a ser manipulada}

\item{vetor_colunas}{Vetor com nome de variáveis adicionais a serem excluídas}
}
\description{
Função que exclui colunas que não continuarão na base final.
Essas variáveis estão sendo excluídas pois não serão necessárias nas análises
das respostas dos instrumentos, sendo importantes apenas no processo de
validação e acompanhamento da aplicação dos instrumentos.
}
\details{
starttime : Horário de abertura do instrumento no aplicativo de aplicação;
endtime : Horário de fechamento do instrumento no aplicativo de aplicação;
uploaddate : Data de aplicação do instrumento no aplicativo de aplicação;
applicationdate : Data da aplicação do instrumento pelos aplicadores;
status : Indica se a observação está válida ou foi deletada;
justification.deletedBy : Indica qual pessoa indicou o deletamento da observação;
justification.reason : Justificativa do deletamento da observação por parte da coordenação do campo de aplicação
}
\examples{
exclui_variaveis()
}
