% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cookbook.R
\name{get_values}
\alias{get_values}
\title{get values}
\usage{
get_values(base_dados, dicionario)
}
\arguments{
\item{base_dados}{Base de dados a ser utilizada}

\item{dicionario}{Dicionario sendo criado}
}
\description{
get values
}
