% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cookbook.R
\name{mode_percentual}
\alias{mode_percentual}
\title{mode percentual}
\usage{
mode_percentual(base_dados, dicionario)
}
\arguments{
\item{base_dados}{Base de dados a ser utilizada}

\item{dicionario}{Dicionario sendo criado}
}
\description{
mode percentual
}
