% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordena_colunas_ceara.R
\name{ordenar_ce_ed}
\alias{ordenar_ce_ed}
\title{Ordena ED Ceará}
\usage{
ordenar_ce_ed(base_ed)
}
\arguments{
\item{base_ed}{Base de dados da Entrevista com Diretores}
}
\description{
Função para reordenar as colunas da Entrevista com Diretores
}
\examples{
ordenar_ed()
}
