% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordena_colunas_ceara.R
\name{ordenar_ce_ep}
\alias{ordenar_ce_ep}
\title{Ordena EP Ceará}
\usage{
ordenar_ce_ep(base_ep)
}
\arguments{
\item{base_ep}{Base de dados da Entrevista com Professores}
}
\description{
Função para reordenar as colunas da Entrevista com Professores
}
\examples{
ordenar_ep()
}
