% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordena_colunas_ceara.R
\name{ordenar_ce_od}
\alias{ordenar_ce_od}
\title{Ordena OD Ceará}
\usage{
ordenar_ce_od(base_od)
}
\arguments{
\item{base_od}{Base de dados da Observação Direta}
}
\description{
Função para reordenar as colunas da Observação Direta
}
\examples{
ordenar_od()
}
