% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reordena_colunas.R
\name{ordenar_ep}
\alias{ordenar_ep}
\title{Ordena EP}
\usage{
ordenar_ep(base_ep)
}
\arguments{
\item{base_ep}{Base de dados da Entrevista com Professores}
}
\description{
Função para reordenar as colunas da Entrevista com Professores
}
\examples{
ordenar_ep()
}
