% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padroniza_nomes.R
\name{padroniza_nomes}
\alias{padroniza_nomes}
\title{Padroniza Nomes}
\usage{
padroniza_nomes(vetor)
}
\arguments{
\item{vetor}{Vetor a ser manipulado}
}
\description{
Função para padronizar os nomes das unidades educacionais e fazer o merge
}
\examples{
padroniza_nomes()
}
