% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padroniza_num.R
\name{padroniza_num}
\alias{padroniza_num}
\title{Padroniza Números}
\usage{
padroniza_num(base)
}
\arguments{
\item{base}{Base a ser manipulada}
}
\description{
Função para padronizar questões numéricas abertas
}
\examples{
padroniza_num()
}
