% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordena_extras.R
\name{realocar_colunas}
\alias{realocar_colunas}
\title{Realocar Colunas}
\usage{
realocar_colunas(base)
}
\arguments{
\item{base}{Base de dados a ser manipulada}
}
\description{
Função que reordena as variáveis extras do instrumento
}
\examples{
realocar_colunas()
}
