% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodifica_multiescolha.R
\name{recodifica_multiescolha}
\alias{recodifica_multiescolha}
\title{Recodifica Multiescolhas}
\usage{
recodifica_multiescolha(variavel, old_cod, new_cod)
}
\arguments{
\item{variavel}{Vetor da variável a ser manipulada}

\item{old_cod}{Código como está na variável}

\item{new_cod}{Código como deve ficar na variável}
}
\description{
Função para recodificar as alternativas nas questões de múltipla escolha
}
\examples{
recodifica_multiescolha()
}
