% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_data.R
\name{resolve_data}
\alias{resolve_data}
\title{Resolve problemas com data}
\usage{
resolve_data(base, colunas)
}
\arguments{
\item{base}{base de dados}

\item{colunas}{Colunas que contém as datas com problema}
}
\value{
Dataframe corrigido
}
\description{
Essa função resolve problemas com a data caso ela venha no formato c('Dia', 'Mes', 'Ano')
}
