% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_extra.R
\name{resolve_extra}
\alias{resolve_extra}
\title{Resolve questões extras com problema}
\usage{
resolve_extra(base, extras)
}
\arguments{
\item{base}{base de dados}

\item{extras}{Colunas que contém as informações com problema}
}
\value{
Dataframe corrigido
}
\description{
Transforma as informações respondidas nas questões extras em um vetor de texto
}
