% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_questoes.R
\name{resolve_questoes}
\alias{resolve_questoes}
\title{Resolve questões que vem dentro de um vetor}
\usage{
resolve_questoes(base, questoes)
}
\arguments{
\item{base}{base de dados}

\item{questoes}{Colunas que contém as informações com problema}
}
\value{
Dataframe corrigido
}
\description{
Essa função resolve problemas com as colunas que vem dentro de um vetor
}
