% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_vetores_concatenados.R
\name{resolve_vetores_concatenados}
\alias{resolve_vetores_concatenados}
\title{Resolve problemas com vetores concatenados}
\usage{
resolve_vetores_concatenados(base, colunas, collapse = "-")
}
\arguments{
\item{base}{base de dados}

\item{colunas}{Colunas que contém as datas com problema}

\item{collapse}{Argumento separador dos valores}
}
\value{
Dataframe corrigido
}
\description{
Essa função resolve problemas com colunas que possuem vetores concatenados.
Um exemplo são os horários de inicio de aplicação de um instrumento. Na base
essas informações estão dispostas em um vetor com valores separados por virgula,
desta forma: "c("11", "30")". O que nós desejamos é deixar com o seguinte formato:
"11:30". Para isso, basta passar o argumento `collapse = ':'` dentro da função.
}
\details{
Também é possível usar a função para resolver problemas com datas. Uma data disposta
dessa maneira: "c("08","02","2021") deverá ficar dessa maneira: "08-02-2021". Para isso,
basta passar o argumento `collapse = '-'`.

Essa função roda independente do número de argumentos em cada vetor.
}
