% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titular_auxiliar.R
\name{titular_auxiliar}
\alias{titular_auxiliar}
\title{Função Titular Auxiliar}
\usage{
titular_auxiliar(base, col, extra)
}
\arguments{
\item{base}{Base de dados a ser manipulada}

\item{col}{Nome da variável, como character, a ser manipulada}

\item{extra}{Vetor de nome das variáveis adicionais, caso seja necessário}
}
\description{
Função para separar a variável referência em análise de professores titulares e auxiliares
}
\examples{
titular_auxiliar()
}
