% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforma_json.R
\name{transforma_json}
\alias{transforma_json}
\title{Transforma o JSON}
\usage{
transforma_json(path, convertList = T)
}
\arguments{
\item{path}{Caminho para o arquivo json}

\item{convertList}{Se a função deverá converter as colunas que são listas em character. Padrão é TRUE.}
}
\value{
Dataframe corrigido
}
\description{
Essa função é responsável por transformar um arquivo formato JSON em
um objeto tipo dataframe.
}
