% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforma_na.R
\name{transforma_na}
\alias{transforma_na}
\title{Transforma NA}
\usage{
transforma_na(base)
}
\arguments{
\item{base}{Base de dados a ser manipulada}
}
\description{
Função para adequar as diversas formas de valores vazios para NA
}
\examples{
transforma_na()
}
