% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforma_vazio.R
\name{transforma_vazio}
\alias{transforma_vazio}
\title{Transforma Vazio}
\usage{
transforma_vazio(base)
}
\arguments{
\item{base}{base de dados a ser utilizada}
}
\description{
Essa função transforma os valores NA's para um character vazio ""
}
\examples{
transforma_vazio()
}
